Rename adformOpenRTBBidAdapter (#3318)
