+++ modules/adformOpenRTBBidAdapter.md
